modified:   UUI-Linux-GTK/App/main.cpp 	new file:   UUI-Linux-GTK/Archive/workingRendererKinda.txt 	modified:   UUI-Linux-GTK/Build/main 	modified:   UUI-Linux-GTK/Include/AngeloCoreFragment.h 	modified:   UUI-Linux-GTK/Source/LinuxGTKRenderer.cpp 	modified:   UniversalUI-Source/CoreRenderer.cpp 	modified:   UniversalUI/Angelo/aRenderCommand.h 	modified:   UniversalUI/Angelo/aRenderOperation.h 	modified:   UniversalUI/Core/uView.h

+++ UniversalUI/Angelo/aRenderOperation.h
@@ -19,6 +19,7 @@
 #include <UniversalUI/Core/CoreGeometry.h>
 #include <UniversalUI/Core/CoreColour.h>
 #include <vector>
+#include <cstdint>
 
 // CODE THKNS  X1    Y1    X2    Y2
 // LINE [2.5, 5.0, 10.0, 10.0, 10.0]
@@ -27,7 +28,7 @@
 // IFL  [0.0, 0.0, 100.0, 50.0]
 
 struct UUI aRenderOperation {
-    unsigned short int code;
+    uint32_t code;
     std::vector<float> data;
 };
